Correct casting for mac
git-svn-id: file:///home/jj/hercules.svn/trunk@6497 956126f8-22a0-4046-8f4a-272fa8102e63

+++ cckddasd.c
@@ -5763,7 +5763,7 @@ int             l;
         if (p)
         {
             l = sprintf ((char *)p, "%s" "." "%6.6ld %1d:%04X> ",
-                tbuf+11, tv.tv_usec, dev ? SSID_TO_LCSS(dev->ssid) : 0, dev ? dev->devnum : 0);
+                tbuf+11, (long)tv.tv_usec, dev ? SSID_TO_LCSS(dev->ssid) : 0, dev ? dev->devnum : 0);
             vsprintf ((char *)p + l, msg, vl);
         }
     }
